Move to common some transaction mappers
