+++ src/routes/transactions/mappers/common/erc20-transfer.mapper.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@nestjs/common';
-import { ModuleTransaction } from '../../../../../domain/safe/entities/module-transaction.entity';
-import { MultisigTransaction } from '../../../../../domain/safe/entities/multisig-transaction.entity';
-import { Token } from '../../../../../domain/tokens/entities/token.entity';
-import { AddressInfoHelper } from '../../../../common/address-info/address-info.helper';
-import { NULL_ADDRESS } from '../../../../common/constants';
-import { TransferTransactionInfo } from '../../../entities/transfer-transaction-info.entity';
-import { Erc20Transfer } from '../../../entities/transfers/erc20-transfer.entity';
+import { ModuleTransaction } from '../../../../domain/safe/entities/module-transaction.entity';
+import { MultisigTransaction } from '../../../../domain/safe/entities/multisig-transaction.entity';
+import { Token } from '../../../../domain/tokens/entities/token.entity';
+import { AddressInfoHelper } from '../../../common/address-info/address-info.helper';
+import { NULL_ADDRESS } from '../../../common/constants';
+import { TransferTransactionInfo } from '../../entities/transfer-transaction-info.entity';
+import { Erc20Transfer } from '../../entities/transfers/erc20-transfer.entity';
 import { DataDecodedParamHelper } from './data-decoded-param.helper';
 import { TransferDirectionHelper } from './transfer-direction.helper';
