+++ src/routes/transactions/mappers/common/native-coin-transfer.mapper.ts
@@ -1,14 +1,14 @@
 import { Injectable } from '@nestjs/common';
-import { ModuleTransaction } from '../../../../../domain/safe/entities/module-transaction.entity';
-import { MultisigTransaction } from '../../../../../domain/safe/entities/multisig-transaction.entity';
-import { Safe } from '../../../../../domain/safe/entities/safe.entity';
-import { AddressInfoHelper } from '../../../../common/address-info/address-info.helper';
-import { AddressInfo } from '../../../../common/entities/address-info.entity';
+import { ModuleTransaction } from '../../../../domain/safe/entities/module-transaction.entity';
+import { MultisigTransaction } from '../../../../domain/safe/entities/multisig-transaction.entity';
+import { Safe } from '../../../../domain/safe/entities/safe.entity';
+import { AddressInfoHelper } from '../../../common/address-info/address-info.helper';
+import { AddressInfo } from '../../../common/entities/address-info.entity';
 import {
   TransferDirection,
   TransferTransactionInfo,
-} from '../../../entities/transfer-transaction-info.entity';
-import { NativeCoinTransfer } from '../../../entities/transfers/native-coin-transfer.entity';
+} from '../../entities/transfer-transaction-info.entity';
+import { NativeCoinTransfer } from '../../entities/transfers/native-coin-transfer.entity';
 
 @Injectable()
 export class NativeCoinTransferMapper {
