+++ src/routes/transactions/mappers/common/settings-change.mapper.spec.ts
@@ -1,18 +1,18 @@
 import { faker } from '@faker-js/faker';
-import { MultisigTransactionBuilder } from '../../../../../domain/safe/entities/__tests__/multisig-transaction.factory';
-import safeFactory from '../../../../../domain/safe/entities/__tests__/safe.factory';
-import { AddressInfoHelper } from '../../../../common/address-info/address-info.helper';
-import { AddressInfo } from '../../../../common/entities/address-info.entity';
-import { AddOwner } from '../../../entities/settings-changes/add-owner.entity';
-import { ChangeMasterCopy } from '../../../entities/settings-changes/change-master-copy.entity';
-import { ChangeThreshold } from '../../../entities/settings-changes/change-threshold.entity';
-import { DeleteGuard } from '../../../entities/settings-changes/delete-guard';
-import { DisableModule } from '../../../entities/settings-changes/disable-module.entity';
-import { EnableModule } from '../../../entities/settings-changes/enable-module.entity';
-import { RemoveOwner } from '../../../entities/settings-changes/remove-owner.entity';
-import { SetFallbackHandler } from '../../../entities/settings-changes/set-fallback-handler.entity';
-import { SetGuard } from '../../../entities/settings-changes/set-guard.entity';
-import { SwapOwner } from '../../../entities/settings-changes/swap-owner.entity';
+import { MultisigTransactionBuilder } from '../../../../domain/safe/entities/__tests__/multisig-transaction.factory';
+import safeFactory from '../../../../domain/safe/entities/__tests__/safe.factory';
+import { AddressInfoHelper } from '../../../common/address-info/address-info.helper';
+import { AddressInfo } from '../../../common/entities/address-info.entity';
+import { AddOwner } from '../../entities/settings-changes/add-owner.entity';
+import { ChangeMasterCopy } from '../../entities/settings-changes/change-master-copy.entity';
+import { ChangeThreshold } from '../../entities/settings-changes/change-threshold.entity';
+import { DeleteGuard } from '../../entities/settings-changes/delete-guard';
+import { DisableModule } from '../../entities/settings-changes/disable-module.entity';
+import { EnableModule } from '../../entities/settings-changes/enable-module.entity';
+import { RemoveOwner } from '../../entities/settings-changes/remove-owner.entity';
+import { SetFallbackHandler } from '../../entities/settings-changes/set-fallback-handler.entity';
+import { SetGuard } from '../../entities/settings-changes/set-guard.entity';
+import { SwapOwner } from '../../entities/settings-changes/swap-owner.entity';
 import { DataDecodedParamHelper } from './data-decoded-param.helper';
 import { SettingsChangeMapper } from './settings-change.mapper';
