+++ src/routes/transactions/mappers/common/settings-change.mapper.ts
@@ -1,21 +1,21 @@
 import { Injectable } from '@nestjs/common';
-import { ModuleTransaction } from '../../../../../domain/safe/entities/module-transaction.entity';
-import { MultisigTransaction } from '../../../../../domain/safe/entities/multisig-transaction.entity';
-import { Safe } from '../../../../../domain/safe/entities/safe.entity';
-import { AddressInfoHelper } from '../../../../common/address-info/address-info.helper';
-import { NULL_ADDRESS } from '../../../../common/constants';
-import { AddressInfo } from '../../../../common/entities/address-info.entity';
-import { AddOwner } from '../../../entities/settings-changes/add-owner.entity';
-import { ChangeMasterCopy } from '../../../entities/settings-changes/change-master-copy.entity';
-import { ChangeThreshold } from '../../../entities/settings-changes/change-threshold.entity';
-import { DeleteGuard } from '../../../entities/settings-changes/delete-guard';
-import { DisableModule } from '../../../entities/settings-changes/disable-module.entity';
-import { EnableModule } from '../../../entities/settings-changes/enable-module.entity';
-import { RemoveOwner } from '../../../entities/settings-changes/remove-owner.entity';
-import { SetFallbackHandler } from '../../../entities/settings-changes/set-fallback-handler.entity';
-import { SetGuard } from '../../../entities/settings-changes/set-guard.entity';
-import { SettingsChange } from '../../../entities/settings-changes/settings-change.entity';
-import { SwapOwner } from '../../../entities/settings-changes/swap-owner.entity';
+import { ModuleTransaction } from '../../../../domain/safe/entities/module-transaction.entity';
+import { MultisigTransaction } from '../../../../domain/safe/entities/multisig-transaction.entity';
+import { Safe } from '../../../../domain/safe/entities/safe.entity';
+import { AddressInfoHelper } from '../../../common/address-info/address-info.helper';
+import { NULL_ADDRESS } from '../../../common/constants';
+import { AddressInfo } from '../../../common/entities/address-info.entity';
+import { AddOwner } from '../../entities/settings-changes/add-owner.entity';
+import { ChangeMasterCopy } from '../../entities/settings-changes/change-master-copy.entity';
+import { ChangeThreshold } from '../../entities/settings-changes/change-threshold.entity';
+import { DeleteGuard } from '../../entities/settings-changes/delete-guard';
+import { DisableModule } from '../../entities/settings-changes/disable-module.entity';
+import { EnableModule } from '../../entities/settings-changes/enable-module.entity';
+import { RemoveOwner } from '../../entities/settings-changes/remove-owner.entity';
+import { SetFallbackHandler } from '../../entities/settings-changes/set-fallback-handler.entity';
+import { SetGuard } from '../../entities/settings-changes/set-guard.entity';
+import { SettingsChange } from '../../entities/settings-changes/settings-change.entity';
+import { SwapOwner } from '../../entities/settings-changes/swap-owner.entity';
 import { DataDecodedParamHelper } from './data-decoded-param.helper';
 
 @Injectable()
