+++ src/routes/transactions/mappers/common/transaction-info.mapper.ts
@@ -1,13 +1,13 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { ModuleTransaction } from '../../../../../domain/safe/entities/module-transaction.entity';
-import { MultisigTransaction } from '../../../../../domain/safe/entities/multisig-transaction.entity';
-import { Operation } from '../../../../../domain/safe/entities/operation.entity';
-import { Safe } from '../../../../../domain/safe/entities/safe.entity';
-import { TokenRepository } from '../../../../../domain/tokens/token.repository';
-import { ITokenRepository } from '../../../../../domain/tokens/token.repository.interface';
-import { TokenType } from '../../../../balances/entities/token-type.entity';
-import { SettingsChangeTransaction } from '../../../entities/settings-change-transaction.entity';
-import { TransactionInfo } from '../../../entities/transaction-info.entity';
+import { ModuleTransaction } from '../../../../domain/safe/entities/module-transaction.entity';
+import { MultisigTransaction } from '../../../../domain/safe/entities/multisig-transaction.entity';
+import { Operation } from '../../../../domain/safe/entities/operation.entity';
+import { Safe } from '../../../../domain/safe/entities/safe.entity';
+import { TokenRepository } from '../../../../domain/tokens/token.repository';
+import { ITokenRepository } from '../../../../domain/tokens/token.repository.interface';
+import { TokenType } from '../../../balances/entities/token-type.entity';
+import { SettingsChangeTransaction } from '../../entities/settings-change-transaction.entity';
+import { TransactionInfo } from '../../entities/transaction-info.entity';
 import { CustomTransactionMapper } from './custom-transaction.mapper';
 import { DataDecodedParamHelper } from './data-decoded-param.helper';
 import { Erc20TransferMapper } from './erc20-transfer.mapper';
